app/requirements.txt
# requirements.txt additions
fastapi[all]==0.104.1
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.4  # For background tasks
redis==5.0.1   # For caching and sessions
boto3==1.34.0  # For AWS S3 file storage
pillow==10.1.0 # For image processing
python-magic==0.4.27  # For file type detection
aiosqlite
filetype